backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sse-starlette>=2.0.0
python-docx>=1.1.0
